Rename files and update references to those files.
